Refactor: ConcreteSessionEnrollment > SessionEnrollment, SessionEnrollment > SessionEnroll 네이밍 수정
