+++ src/main/java/nextstep/courses/domain/enrollment/PaidSessionEnrollment.java
@@ -1,10 +1,10 @@
 package nextstep.courses.domain.enrollment;
 
-import nextstep.courses.domain.enrollment.engine.ConcreteSessionEnrollment;
+import nextstep.courses.domain.enrollment.engine.SessionEnrollment;
 import nextstep.courses.exception.SessionFeeMismatchException;
 import nextstep.payments.domain.Payment;
 
-public class PaidSessionEnrollment extends ConcreteSessionEnrollment {
+public class PaidSessionEnrollment extends SessionEnrollment {
 
     public PaidSessionEnrollment(Long sessionId, SessionStatus status, int capacity, long fee) {
         super(sessionId, status, capacity, fee);
