+++ src/main/java/nextstep/courses/domain/enrollment/engine/SessionEnrollment.java
@@ -1,16 +1,55 @@
 package nextstep.courses.domain.enrollment.engine;
 
+import nextstep.courses.domain.enrollment.SessionCapacity;
+import nextstep.courses.domain.enrollment.SessionFee;
+import nextstep.courses.domain.enrollment.SessionStatus;
+import nextstep.courses.domain.enrollment.SessionStudent;
+import nextstep.courses.exception.SessionCapacityExceedException;
 import nextstep.payments.domain.Payment;
 import nextstep.users.domain.NsUser;
 
-public interface SessionEnrollment {
+import java.util.ArrayList;
+import java.util.List;
 
-    void enroll(NsUser nsUser, Payment payment);
+public abstract class SessionEnrollment implements SessionEnroll {
 
-    default void satisfyEnrollment(Payment payment) {
-        satisfyStatus();
-        satisfyCapacity();
-        satisfyFee(payment);
+    protected Long sessionId;
+    protected final SessionStatus status;
+    protected final SessionCapacity capacity;
+    protected final SessionFee fee;
+    protected final List<SessionStudent> students;
+
+    protected SessionEnrollment(Long sessionId, SessionStatus status, int capacity, long fee) {
+        this.sessionId = sessionId;
+        this.status = status;
+        this.capacity = new SessionCapacity(sessionId, capacity);
+        this.fee = new SessionFee(sessionId, fee);
+        this.students = new ArrayList<>();
+    }
+
+    protected SessionEnrollment(SessionStatus status, int capacity, long fee) {
+        this.status = status;
+        this.capacity = new SessionCapacity(capacity);
+        this.fee = new SessionFee(fee);
+        this.students = new ArrayList<>();
+    }
+
+    @Override
+    public void enroll(NsUser nsUser, Payment payment) {
+        satisfyEnrollment(payment);
+        students.add(SessionStudent.from(sessionId, nsUser));
+    }
+
+    @Override
+    public void satisfyStatus() {
+        status.canEnroll();
+    }
+
+    @Override
+    public void satisfyCapacity() {
+        if (capacity.noCapacity(students.size())) {
+            throw new SessionCapacityExceedException(capacity.get(), students.size());
+        }
     }
 
     void satisfyStatus();
